restructure components into subdirectories
